+++ server/app/uk/gov/ons/addressIndex/server/controllers/AddressController.scala
@@ -5,7 +5,6 @@ import play.api.libs.json.Json
 import play.api.mvc._
 import uk.gov.ons.addressIndex.model.db.index.HybridAddresses
 import uk.gov.ons.addressIndex.model.server.response._
-import uk.gov.ons.addressIndex.parsers.Tokens
 import uk.gov.ons.addressIndex.server.modules.response.AddressIndexResponse
 import uk.gov.ons.addressIndex.server.modules.validation.AddressValidation
 import uk.gov.ons.addressIndex.server.modules.{ConfigModule, ElasticsearchRepository, ParserModule, VersionModule}
@@ -33,95 +32,6 @@ class AddressController @Inject()(val controllerComponents: ControllerComponents
   val invalid: String = "invalid"
   val valid: String = "valid"
   val notRequired: String = "not required"
-  val DATE_FORMAT = "yyyy-MM-dd"
-
-  /**
-    * Codelist List API
-    *
-    * @return Json response with codelist
-    */
-  def codeList(): Action[AnyContent] = Action async { implicit req =>
-    val codList = Tokens.codeList.map { clval =>
-
-      new AddressResponseCodelist(name = clval.split("=").headOption.getOrElse(""),
-        description = clval.split("=").lastOption.getOrElse(""))
-    }
-
-    val codeListContainer = new AddressResponseCodelistListContainer(codList)
-
-    Future(Ok(Json.toJson(codeListContainer)))
-  }
-
-  /**
-    * Classification List API
-    *
-    * @return Json response with codelist
-    */
-  def codeListClassification(): Action[AnyContent] = Action async { implicit req =>
-    val classList = Tokens.classList.map { classval =>
-
-      new AddressResponseClassification(code = classval.split("=").headOption.getOrElse(""),
-        label = classval.split("=").lastOption.getOrElse(""))
-    }
-
-    val codListContainer = new AddressResponseClassificationListContainer(classList)
-    Future(Ok(Json.toJson(codListContainer)))
-  }
-
-  /**
-    * Custodian List API
-    *
-    * @return Json response with codelist
-    */
-  def codeListCustodian(): Action[AnyContent] = Action async { implicit req =>
-    val custList = Tokens.custodianList.map { custval =>
-
-      new AddressResponseCustodian(
-        custval.split(",").lift(0).getOrElse(""),
-        custval.split(",").lift(1).getOrElse(""),
-        custval.split(",").lift(2).getOrElse(""),
-        custval.split(",").lift(3).getOrElse(""),
-        custval.split(",").lift(4).getOrElse(""),
-        custval.split(",").lift(5).getOrElse("")
-      )
-    }
-
-    val custListContainer = new AddressResponseCustodianListContainer(custList)
-    Future(Ok(Json.toJson(custListContainer)))
-  }
-
-  /**
-    * Source List API
-    *
-    * @return Json response with codelist
-    */
-  def codeListSource(): Action[AnyContent] = Action async { implicit req =>
-
-    val sourceList = Tokens.sourceList.map { sourceval =>
-      new AddressResponseSource(sourceval.split("=").headOption.getOrElse(""),
-        sourceval.split("=").lastOption.getOrElse(""))
-    }
-
-    val sourceListContainer = new AddressResponseSourceListContainer(sourceList)
-    Future(Ok(Json.toJson(sourceListContainer)))
-  }
-
-  /**
-    * Logical Status List API
-    *
-    * @return Json response with codelist
-    */
-  def codeListLogicalStatus(): Action[AnyContent] = Action async { implicit req =>
-
-    val logicalList = Tokens.logicalStatusList.map { logstatval =>
-      new AddressResponseLogicalStatus(logstatval.split("=").headOption.getOrElse(""),
-        logstatval.split("=").lastOption.getOrElse(""))
-    }
-
-    val logicalListContainer = new AddressResponseLogicalStatusListContainer(logicalList)
-    Future(Ok(Json.toJson(logicalListContainer)))
-  }
-
 
   /**
     * Address query API
@@ -289,129 +199,6 @@ class AddressController @Inject()(val controllerComponents: ControllerComponents
                 writeLog(badRequestErrorMessage = FailedRequestToEsError.message)
                 logger.warn(s"Could not handle individual request (address input), problem with ES ${exception.getMessage}")
                 InternalServerError(Json.toJson(FailedRequestToEs))
-            }
-        }
-    }
-  }
-
-  /**
-    * PartialAddress query API
-    *
-    * @param input input for the address to be fetched
-    * @return Json response with addresses information
-    */
-
-  def partialAddressQuery(input: String, offset: Option[String] = None, limit: Option[String] = None,
-                          classificationfilter: Option[String] = None, startDate: Option[String], endDate: Option[String],
-			  historical: Option[String] = None): Action[AnyContent] = Action async { implicit req =>
-    val startingTime = System.currentTimeMillis()
-
-    // get the defaults and maxima for the paging parameters from the config
-    val defLimit = conf.config.elasticSearch.defaultLimitPartial
-    val defOffset = conf.config.elasticSearch.defaultOffset
-
-    val limval = limit.getOrElse(defLimit.toString)
-    val offval = offset.getOrElse(defOffset.toString)
-
-    val filterString = classificationfilter.getOrElse("")
-
-    val startDateVal = startDate.getOrElse("")
-    val endDateVal = endDate.getOrElse("")
-
-    val hist = historical match {
-      case Some(x) => Try(x.toBoolean).getOrElse(true)
-      case None => true
-    }
-
-    def writeLog(doResponseTime: Boolean = true, badRequestErrorMessage: String = "", notFound: Boolean = false, formattedOutput: String = "", numOfResults: String = "", score: String = ""): Unit = {
-      val responseTime = if (doResponseTime) (System.currentTimeMillis() - startingTime).toString else ""
-      val networkid = req.headers.get("authorization").getOrElse("Anon").split("_")(0)
-      logger.systemLog(
-        ip = req.remoteAddress, url = req.uri, responseTimeMillis = responseTime,
-        partialAddress = input, isNotFound = notFound, offset = offval,
-        limit = limval, filter = filterString, badRequestMessage = badRequestErrorMessage,
-        formattedOutput = formattedOutput,
-        numOfResults = numOfResults, score = score, networkid = networkid, startDate = startDateVal, endDate = endDateVal,
-        historical = hist
-      )
-    }
-
-    val limitInt = Try(limval.toInt).toOption.getOrElse(defLimit)
-    val offsetInt = Try(offval.toInt).toOption.getOrElse(defOffset)
-
-    val result: Option[Future[Result]] =
-      addressValidation.validateAddressLimit(limit)
-        .orElse(addressValidation.validateAddressOffset(offset))
-        .orElse(addressValidation.validateStartDate(startDateVal))
-        .orElse(addressValidation.validateEndDate(endDateVal))
-        .orElse(addressValidation.validateSource)
-        .orElse(addressValidation.validateKeyStatus)
-        .orElse(addressValidation.validateInput(input))
-        .orElse(None)
-
-    result match {
-
-      case Some(res) =>
-        res // a validation error
-
-      case _ =>
-
-        val request: Future[HybridAddresses] =
-          overloadProtection.breaker.withCircuitBreaker(
-            esRepo.queryPartialAddress(input, offsetInt, limitInt, filterString, startDateVal, endDateVal, None, hist)
-          )
-
-        request.map {
-          case HybridAddresses(hybridAddresses, maxScore, total) =>
-            //TODO: Verify AddressResponsePartialAddress is correct after the merge
-            val addresses: Seq[AddressResponsePartialAddress] = hybridAddresses.map(
-              AddressResponsePartialAddress.fromHybridAddress
-            )
-
-            addresses.foreach { address =>
-              writeLog(
-                formattedOutput = address.formattedAddressNag, numOfResults = total.toString,
-                score = address.underlyingScore.toString
-              )
-            }
-
-            writeLog()
-
-            jsonOk(
-              AddressByPartialAddressResponseContainer(
-                apiVersion = apiVersion,
-                dataVersion = dataVersion,
-                response = AddressByPartialAddressResponse(
-                  input = input,
-                  addresses = addresses,
-                  filter = filterString,
-                  historical = hist,
-                  limit = limitInt,
-                  offset = offsetInt,
-                  total = total,
-                  maxScore = maxScore,
-                  startDate = startDateVal,
-                  endDate = endDateVal
-                ),
-                status = OkAddressResponseStatus
-              )
-            )
-
-        }.recover {
-          case NonFatal(exception) =>
-
-            overloadProtection.currentStatus match {
-              case ThrottlerStatus.HalfOpen =>
-                logger.warn(s"Elasticsearch is overloaded or down (address input). Circuit breaker is Half Open: ${exception.getMessage}")
-                TooManyRequests(Json.toJson(FailedRequestToEsTooBusy))
-              case ThrottlerStatus.Open =>
-                logger.warn(s"Elasticsearch is overloaded or down (address input). Circuit breaker is open: ${exception.getMessage}")
-                TooManyRequests(Json.toJson(FailedRequestToEsTooBusy))
-              case _ =>
-                // Circuit Breaker is closed. Some other problem
-                writeLog(badRequestErrorMessage = FailedRequestToEsPartialAddressError.message)
-                logger.warn(s"Could not handle individual request (partialAddress input), problem with ES ${exception.getMessage}")
-                InternalServerError(Json.toJson(FailedRequestToEsPartialAddress))
             }
         }
     }
